Move image loading and saving utils to /utils
PiperOrigin-RevId: 438403293
