+++ keras/utils/__init__.py
@@ -26,6 +26,12 @@ from keras.utils.vis_utils import plot_model
 from keras.utils.np_utils import normalize
 from keras.utils.np_utils import to_categorical
 
+# Image related
+from keras.utils.image_utils import array_to_img
+from keras.utils.image_utils import img_to_array
+from keras.utils.image_utils import load_img
+from keras.utils.image_utils import save_img
+
 # Sequence related
 from keras.utils.data_utils import Sequence
 from keras.utils.data_utils import GeneratorEnqueuer
